pyproject.toml
[build-system]
requires = ["setuptools>=72"]
build-backend = "setuptools.build_meta"

[project]
name = "data_analysis_package"
version = "0.0.1"
description = "A package for data analysis"
authors = [
    { name = "Anna Czartoszewska", email = "a.czartoszewsk@student.uw.edu.pl"}
]
#classifiers = [
#    "Development Status :: 3 - Alpha",
#    "Intended Audience :: Developers",
#    "License :: OSI Approved :: BSD License",
#    "Programming Language :: Python :: 3",
#    "Programming Language :: Python :: 3.10",
#    "Programming Language :: Python :: 3.11",
#    "Programming Language :: Python :: 3.12",
#    "Programming Language :: Python :: 3.13",
#]
dynamic=["readme"]
dependencies = [
    "pandas >= 1.3.0",
    "lxml",
    "xlrd >= 2.0.1",
    "openpyxl",
    "matplotlib.pyplot",
    "scipy.stats",
    "seaborn"
]


[tool.setuptools.packages.find]
namespaces = false

[tool.setuptools.dynamic]
readme = {file = "Readme.rst"}
